+++ examples/api-platform/src/layouts/Admin.vue
@@ -5,8 +5,9 @@
       :header-menu="headerMenu"
       dense
       dark
-      @drawer="drawer = !drawer"
-      @mini-variant="mini = !mini"
+      @toggle="
+        $vuetify.breakpoint.lgAndUp ? (mini = !mini) : (drawer = !drawer)
+      "
     ></va-app-bar>
     <va-sidebar
       slot="sidebar"