+++ examples/demo/src/layouts/Admin.vue
@@ -7,8 +7,9 @@
       color="secondary"
       dense
       dark
-      @drawer="drawer = !drawer"
-      @mini-variant="mini = !mini"
+      @toggle="
+        $vuetify.breakpoint.lgAndUp ? (mini = !mini) : (drawer = !drawer)
+      "
     ></va-app-bar>
     <va-sidebar
       slot="sidebar"