+++ packages/cli/generator/admin/src/layouts/Admin.vue
@@ -6,8 +6,9 @@
       :profile-menu="profileMenu"
       dense
       dark
-      @drawer="drawer = !drawer"
-      @mini-variant="mini = !mini"
+      @toggle="
+        $vuetify.breakpoint.lgAndUp ? (mini = !mini) : (drawer = !drawer)
+      "
     ></va-app-bar>
     <va-sidebar
       slot="sidebar"